requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
idna==3.4
lxml==4.9.1
psycopg2-binary==2.9.5
requests==2.28.1
urllib3==1.26.12
